begin.el org loading mechanism altered
It now only loads org if tangling is needed, and unloads org once that
is done. Moreover file freshness checks, tangling, and compiling are
all done as bulk operations now. Previously each file was checked,
tangled and compiled before the next file would be considered.

As a result, several macros in my-load-macros were rendered moot and
have been removed.

Additionally, some changes had to be made to
init/lit-emacs-init-org.org to compensate for the above. Some
keybindings that were previously defined in begin.el have moved
here.

And, on a slightly unrelated note, the version check on org that
happens in init/lit-emacs-init-org.org is now run as part of :config
rather than :preface. This change was needed now that org is no longer
unconditionally required in begin.el.